Create Data call module
